OWNERS: Add Manoj as an owner
The bios-setting-mgr repository has been in maintenance mode for almost
a couple of years now, and there seems to be a lack of active reviewers
and maintainers. Given that we do not anticipate much active
development in this repo, I would be happy to step in as a maintainer
and ensure that we adhere to upstream coding, C++, and clang standards,
keeping the repository in alignment with the latest practices. I look
forward to helping with the ongoing maintenance and ensuring smooth
support.

Change-Id: I32874939a453b948a6375443629c2f29b644e3c7
Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>