Refactor the biosData loading version
- Move credentialBootstrapFlag loading to version 3 only.
- Update BIOS data persistence file changes:
  1. Add a pending boot option in 24.04 release.
  2. Add a string to the bioTable in 24.05 release,
     and save the version at the beginning.
  3. Save credentialBootstrapFlag in 24.10 release.

- Remove try-catch block in the load function to handle
 cases where credentialBootstrapFlag is expected but
 not present in the file, which previously caused
 errors when trying to load bootOption variant type.
Now all the errors will be caught in deserialize.

- No changes are needed for pending boot order data as
it is a map of variants, and no new type was added
to the variant. During restart of the service,
the pending enable value will be initialized
from the enable value, otherwise in the case of version
change it may take a random value.

Testing was done using biosData file from older version
Fixes nvbug https://redmine.mellanox.com/issues/4188647

Signed-off-by: afogel <afogel@afogel2.nvidia.com>